Refactor multiple PHP scripts for improved readability and security; streamline variable definitions, enhance query handling, and implement consistent error handling.

+++ src/emptytrash.php
@@ -1,25 +1,15 @@
 <?php
 	date_default_timezone_set('UTC');
-	include 'functions.php';
-    require 'config.php';
+	require 'config.php';
 	include 'db_connect.php';
     
-    $query1 = 'SELECT id FROM entries WHERE trash = 1';
-    $res = $con->query($query1);
-    while($row = mysqli_fetch_array($res, MYSQLI_ASSOC))
-    {
+    // Delete all files from trash entries
+    $res = $con->query('SELECT id FROM entries WHERE trash = 1');
+    while($row = mysqli_fetch_array($res, MYSQLI_ASSOC)) {
         $file_path = "entries/".$row["id"].".html";
-        if(file_exists($file_path))
-        {
-            unlink($file_path);
-        }
-        else
-        {
-            echo 'file not found';
-        }
+        if(file_exists($file_path)) unlink($file_path);
     }
     
-	$query2 = "DELETE from entries WHERE trash = 1";
-	if($con->query($query2)) echo 1;
-	else echo mysqli_error($con);
+    // Delete all trash entries
+	echo $con->query("DELETE FROM entries WHERE trash = 1") ? 1 : mysqli_error($con);
 ?>
